refactors state/useEffect for night mode and emoji toggle, changes coloring on landing page
