+++ src/Components/NavBars/SecondaryNav/SecondaryNav.jsx
@@ -5,7 +5,7 @@ import Logo from '../MandalaLogo/Logo.png';
 import '../LandingNav/LandingNav.css';
 import './SecondaryNav.css';
 
-export default function SecondaryNav() {
+export default function SecondaryNav({darkMode, setDarkMode, emoji, setEmoji}) {
     return (
         <Navbar bg="light" variant="light">
 			<Container>
@@ -14,7 +14,13 @@ export default function SecondaryNav() {
 			<Nav.Link href="" className="navLink"><Link to="/">ANIMALS</Link></Nav.Link>
 			<Nav.Link href="" className="navLink"><Link to="">HOLIDAYS</Link></Nav.Link>
 			<Nav.Link href="" className="navLink"><Link to="">MANDALAS</Link></Nav.Link>
-			<Nav.Link><button className="dark-btn">🌙 </button></Nav.Link>
+			<Nav.Link>
+				<button className="dark-btn" 
+						onClick={() => {setEmoji(!emoji); setDarkMode(!darkMode)}}>
+								{ emoji ? "☀️  " : " 🌙   " }	
+						</button>
+							{emoji}
+			</Nav.Link>
 			</Nav>
 			</Container>
 		</Navbar>
