+++ src/Pages/App/App.css
@@ -1,16 +1,17 @@
 @import url('https://fonts.googleapis.com/css2?family=Rubik:wght@600&family=Urbanist:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
 
 .dark {
-  background: #36393E;
+  background: #303030;
+
 }
 
 .dark p, h2, h3, h5 {
   color: var(--white);
 }
 
-.dark img.l-page-icons {
+/* .dark img.l-page-icons {
   background-color: rgba(255,255,255, 0.5);
   border-radius: 100px;
   filter: grayscale(50%) opacity(90%);
-}
+} */
