+++ src/Pages/ColorBook/ColorBook.jsx
@@ -5,10 +5,10 @@ import PaintBox from '../../Components/PaintBox/PaintBox';
 import './ColorBook.css';
 
 
-export default function ColorBook() {
+export default function ColorBook({darkMode, setDarkMode, emoji, setEmoji}) {
     return (
         <div className="colorbook">
-            <SecondaryNav />
+            <SecondaryNav darkMode={darkMode} setDarkMode={setDarkMode} emoji={emoji} setEmoji={setEmoji}/>
             <PaintBox />
         </div>
     )
